Refactor cdrm java classes to ucar.gcdm package.
