+++ .run/TestCdmrNetcdfFile.run.xml
@@ -4,12 +4,12 @@
     <useClassPathOnly />
     <extension name="coverage">
       <pattern>
-        <option name="PATTERN" value="ucar.cdmr.client.*" />
+        <option name="PATTERN" value="ucar.gcdm.client.*" />
         <option name="ENABLED" value="true" />
       </pattern>
     </extension>
-    <option name="PACKAGE_NAME" value="ucar.cdmr.client" />
-    <option name="MAIN_CLASS_NAME" value="ucar.cdmr.client.TestCdmrNetcdfFile" />
+    <option name="PACKAGE_NAME" value="ucar.gcdm.client" />
+    <option name="MAIN_CLASS_NAME" value="ucar.gcdm.TestGcdmNetcdfFile" />
     <option name="METHOD_NAME" value="" />
     <option name="TEST_OBJECT" value="class" />
     <option name="VM_PARAMETERS" value="-ea -Dunidata.testdata.path=/media/snake/0B681ADF0B681ADF/thredds-test-data/local/thredds-test-data/ -Djava.util.logging.config.file=src/main/resources/logging.cfg" />
