--- cdmr/src/main/proto/ucar/cdmr/cdmrServer.proto
@@ -1,15 +1,0 @@
-syntax = "proto3";
-
-import "ucar/cdmr/cdmrNetcdf.proto";
-import "ucar/cdmr/cdmrGrid.proto";
-
-package ucar.cdmr;
-option java_package = "ucar.cdmr";
-option java_outer_classname = "CdmrServerProto";
-
-service CdmRemote {
-  rpc GetNetcdfHeader (HeaderRequest) returns (HeaderResponse) {}
-  rpc GetNetcdfData (DataRequest) returns (stream DataResponse) {}
-  rpc GetGridDataset (GridDatasetRequest) returns (GridDatasetResponse) {}
-  rpc GetGridData (GridDataRequest) returns (stream GridDataResponse) {}
-}
