--- cdmr/src/main/resources/META-INF/services/ucar.nc2.dataset.spi.NetcdfFileProvider
@@ -1,1 +1,0 @@
-ucar.cdmr.client.CdmrNetcdfFileProvider
