+++ gcdm/build.gradle
@@ -1,0 +1,77 @@
+import com.github.psxpaul.task.JavaExecFork
+
+plugins {
+  // Provide convenience executables for testing
+  id 'application'
+  id "com.google.protobuf"
+  id 'com.github.psxpaul.execfork' version '0.1.13'
+}
+
+description = 'gRPC client and server implementation of CDM Remote Procedure Calls (gCDM).'
+ext.title = 'CDM Remote Procedure Calls'
+
+apply from: "$rootDir/gradle/any/dependencies.gradle"
+apply from: "$rootDir/gradle/any/java-library.gradle"
+apply from: "$rootDir/gradle/any/protobuf.gradle"
+
+dependencies {
+  api project(':cdm-core')
+
+  // for CdmrServer: we should just add on the classpath?
+  api project(':bufr')
+  api project(':grib')
+
+  implementation "io.grpc:grpc-protobuf:${depVersion.grpc}"
+  implementation "io.grpc:grpc-stub:${depVersion.grpc}"
+  compileOnly 'org.apache.tomcat:annotations-api:6.0.53'
+
+  runtimeOnly "io.grpc:grpc-netty-shaded:${depVersion.grpc}"
+  runtimeOnly "org.slf4j:slf4j-jdk14"
+
+  testRuntimeOnly "org.slf4j:slf4j-jdk14:${depVersion.slf4j}"
+  testImplementation enforcedPlatform(project(':netcdf-java-testing-platform'))
+  testImplementation project(':cdm-test-utils')
+  testImplementation 'com.google.truth:truth'
+  testImplementation 'junit:junit'
+  testImplementation 'commons-io:commons-io'
+  testImplementation "org.mockito:mockito-core"
+  testImplementation "io.grpc:grpc-testing:${depVersion.grpc}"
+}
+
+protobuf {
+  protoc { artifact = "com.google.protobuf:protoc:${depVersion.protobuf}" }
+  plugins {
+    grpc { artifact = "io.grpc:protoc-gen-grpc-java:${depVersion.grpc}" }
+  }
+  generateProtoTasks {
+    all()*.plugins { grpc {} }
+  }
+}
+
+application {
+  mainClass = 'ucar.gcdm.server.GcdmServer'
+}
+
+// Inform IDEs like IntelliJ IDEA, Eclipse or NetBeans about the generated code.
+sourceSets {
+  main {
+    java {
+      // not needed for proto generated source
+      srcDirs 'build/generated/source/proto/main/grpc'
+      srcDirs 'build/generated/source/proto/main/java'
+    }
+  }
+}
+
+task startDaemon(type: JavaExecFork) {
+  classpath = sourceSets.main.runtimeClasspath
+  main = 'ucar.gcdm.server.GcdmServer'
+  jvmArgs = [ '-Xmx512m', '-Djava.awt.headless=true' ]
+  standardOutput = "$buildDir/gcdm_logs/gcdm.log"
+  errorOutput = "$buildDir/gcdm_logs/gcdm-error.log"
+  stopAfter = test
+  waitForPort = 16111
+  waitForOutput = 'Server started, listening on 16111'
+}
+
+test.dependsOn(startDaemon)
