+++ gcdm/src/main/java/ucar/gcdm/client/GcdmGridDatasetProvider.java
@@ -1,0 +1,25 @@
+package ucar.gcdm.client;
+
+import thredds.client.catalog.ServiceType;
+import ucar.nc2.dataset.DatasetUrl;
+import ucar.nc2.dataset.spi.GridDatasetProvider;
+import ucar.nc2.grid.GridDataset;
+import ucar.nc2.util.CancelTask;
+
+public class GcdmGridDatasetProvider implements GridDatasetProvider {
+  @Override
+  public String getProtocol() {
+    return "cdmr";
+  }
+
+  @Override
+  public boolean isOwnerOf(DatasetUrl url) {
+    return url.getServiceType() == ServiceType.Cdmr;
+  }
+
+  @Override
+  public GridDataset open(String location, CancelTask cancelTask) {
+    return GcdmGridDataset.builder().setRemoteURI(location).build();
+  }
+
+}
