+++ gcdm/src/main/proto/ucar/gcdm/gcdmServer.proto
@@ -1,0 +1,15 @@
+syntax = "proto3";
+
+import "ucar/gcdm/cdmrNetcdf.proto";
+import "ucar/gcdm/cdmrGrid.proto";
+
+package ucar.cdmr;
+option java_package = "ucar.cdmr";
+option java_outer_classname = "CdmrServerProto";
+
+service CdmRemote {
+  rpc GetNetcdfHeader (HeaderRequest) returns (HeaderResponse) {}
+  rpc GetNetcdfData (DataRequest) returns (stream DataResponse) {}
+  rpc GetGridDataset (GridDatasetRequest) returns (GridDatasetResponse) {}
+  rpc GetGridData (GridDataRequest) returns (stream GridDataResponse) {}
+}
