+++ gcdm/src/main/resources/META-INF/services/ucar.nc2.dataset.spi.GridDatasetProvider
@@ -1,0 +1,1 @@
+ucar.gcdm.client.GcdmGridDatasetProvider
