+++ gcdm/src/main/resources/logging.cfg
